+++ src/index.ts
@@ -2,3 +2,5 @@ export * from "./GameClient";
 export * from "./GameContext";
 export * from "./loggers";
 export * from "./types";
+
+export * from "@akashic/headless-driver";