+++ src/types.ts
@@ -1,3 +1,16 @@
-import { RunnerV1Game, RunnerV2Game, RunnerV3Game } from "@akashic/headless-driver";
+import { RunnerV1Game, RunnerV1_g, RunnerV2Game, RunnerV2_g, RunnerV3Game, RunnerV3_g } from "@akashic/headless-driver";
 
-export type RunnerGame = RunnerV1Game | RunnerV2Game | RunnerV3Game;
+export interface EngineVersions {
+	1: {
+		g: RunnerV1_g;
+		game: RunnerV1Game;
+	};
+	2: {
+		g: RunnerV2_g;
+		game: RunnerV2Game;
+	};
+	3: {
+		g: RunnerV3_g;
+		game: RunnerV3Game;
+	};
+}