+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 2.0.0
+* headless-driver@1.7.0 に追従
+* `GameClientStartParameterObject#renderingMode` を追加
+* `GameClient#getPrimarySurfaceCanvas()` を追加
+* `GameClient#g` を追加
+* `GameClient`, `GameClientParameterObject` の generics の型を変更
+
 ## 1.1.1
 * `GameContextParameterObject#gameJsonPath` を省略した場合に空のゲームコンテンツを実行できるように